+++ CHANGES.md
@@ -80,6 +80,10 @@ v0.14.3 (unreleased)
 * Fix bug that caused an error when using the profile collapse tool and
   dragging from right to left. [#2002]
 
+* Fix bug that caused labels on the x-axis of the histogram viewer to
+  be incorrectly set to numbers instead of strings when showing a
+  histogram of a string variable and saving/reloading session. [#2009]
+
 v0.14.2 (2019-02-04)
 --------------------